Fixed a movehistorydouble bug, probably after merge. Score now takes a percentage of the score into account as well.

+++ Console4SimpleStorage/Console4SimpleStorage/Node.h
@@ -21,7 +21,7 @@ public:
 	bool isRoot;
 	double value;
 	double probability;
-	int sum;
+	double sum;
 	Option opt;
 	std::vector<Node*> childrenPtrs;
 	Node* parentPtr;
